refactor: decompose pipeline.rs into submodules
Extract src/pipeline.rs (508 lines) into src/pipeline/ with 7 focused
submodules: mod.rs (re-exports, types), cache.rs (thread-local
compilation cache), scan.rs (pre-scanning with unified prescan_forms),
fixpoint.rs (shared fixpoint loop), compile.rs, analyze.rs, eval.rs.

Key improvement: the fixpoint effect-inference loop was duplicated
between compile_all and analyze_all — now unified in run_fixpoint()
parameterized by a post_analyze closure (compile passes mark_tail_calls,
analyze passes a no-op). Pre-scanning consolidated from two separate
passes into one via prescan_forms().

No behavioral changes. All public API signatures preserved via
re-exports in mod.rs. All existing tests pass.

Update docs/pipeline.md to reflect new module structure.

Fixes #375
